# Accessibility and performance fixes