--- a/public/_headers
+++ b/public/_headers
@@ -5,3 +5,6 @@
   cache-control: max-age=31536000
   cache-control: immutable
   cache-control: public
+/fonts/*
+  cache-control: max-age=604800
+  cache-control: public