--- a/src/helpers/path.ts
+++ b/src/helpers/path.ts
@@ -1,0 +1,2 @@
+/** Adds a trailing `/` to the passed path if needed. */
+export const ensureTrailingSlash = (path: string) => (path.at(-1) === '/' ? path : `${path}/`);