--- a/src/icons/logos/next-js.svg
+++ b/src/icons/logos/next-js.svg
@@ -1,0 +1,15 @@
+<svg xmlns="http://www.w3.org/2000/svg" fill="none" viewBox="0 0 20 20">
+  <path stroke="#fff" stroke-width="1.5" d="M19.25 10a9.25 9.25 0 1 1-18.5 0 9.25 9.25 0 0 1 18.5 0Z"/>
+  <path fill="url(#a)" d="M16.61 17.5 7.68 6H6v8h1.35V7.7l8.2 10.62c.38-.25.73-.52 1.06-.82Z"/>
+  <path fill="url(#b)" d="M14.11 6h-1.33v8h1.33V6Z"/>
+  <defs>
+    <linearGradient id="a" x1="12.11" x2="16.06" y1="12.94" y2="17.83" gradientUnits="userSpaceOnUse">
+      <stop stop-color="#fff"/>
+      <stop offset="1" stop-color="#fff" stop-opacity="0"/>
+    </linearGradient>
+    <linearGradient id="b" x1="13.44" x2="13.42" y1="6" y2="11.88" gradientUnits="userSpaceOnUse">
+      <stop stop-color="#fff"/>
+      <stop offset="1" stop-color="#fff" stop-opacity="0"/>
+    </linearGradient>
+  </defs>
+</svg>