--- a/src/pages/_components/landing-page/Ecosystem.astro
+++ b/src/pages/_components/landing-page/Ecosystem.astro
@@ -39,7 +39,7 @@ import Trending from './ecosystem-tabs/Trending.astro';
 
 	<div class="my-4 w-full border-t border-t-astro-gray-600"></div>
 
-	<ul class="mx-auto w-full max-w-screen-xl px-4 sm:px-8">
+	<div class="mx-auto w-full max-w-screen-xl px-4 sm:px-8">
 		<EcosystemCategory id="1" tab="trending">
 			<Trending />
 		</EcosystemCategory>
@@ -58,7 +58,7 @@ import Trending from './ecosystem-tabs/Trending.astro';
 		<EcosystemCategory id="6" tab="landing pages" class="hidden">
 			<LandingPages />
 		</EcosystemCategory>
-	</ul>
+	</div>
 </section>
 
 <script>