--- a/src/pages/_components/landing-page/EcosystemCategory.astro
+++ b/src/pages/_components/landing-page/EcosystemCategory.astro
@@ -8,7 +8,7 @@ type Props = {
 const { id, tab, class: className } = Astro.props;
 ---
 
-<li
+<section
 	data-tab={tab}
 	role="tabpanel"
 	aria-labelledby={tab}
@@ -16,4 +16,4 @@ const { id, tab, class: className } = Astro.props;
 	id={`ecosystem-panel-${id}`}
 >
 	<slot />
-</li>
+</section>