--- a/src/pages/_components/landing-page/HeroBackground.astro
+++ b/src/pages/_components/landing-page/HeroBackground.astro
@@ -7,6 +7,6 @@ const background = await getImage({ src: heroBackground, quality: 10, width: 256
 
 <div
 	style={`background-image: url('${background.src}');`}
-	class="blur-3xl absolute -z-50 m-auto -inset-1/2 bg-[length:2353px_1969px] w-[2353px] h-[1969px]"
+	class="blur-lg absolute -z-50 m-auto -inset-1/2 bg-[length:2353px_1969px] w-[2353px] h-[1969px]"
 >
 </div>