--- a/src/pages/_components/landing-page/IntegrationContent.astro
+++ b/src/pages/_components/landing-page/IntegrationContent.astro
@@ -8,7 +8,7 @@ interface Props {
 const { id, tab, class: className } = Astro.props;
 ---
 
-<li
+<section
 	data-tab={tab}
 	role="tabpanel"
 	aria-labelledby={tab}
@@ -38,4 +38,4 @@ const { id, tab, class: className } = Astro.props;
 			<slot name="example" />
 		</div>
 	</div>
-</li>
+</section>