--- a/src/pages/_components/landing-page/IntegrationTab.astro
+++ b/src/pages/_components/landing-page/IntegrationTab.astro
@@ -39,7 +39,7 @@ const {
 					innerBackground,
 				]}
 			>
-				<img class={size} src={logoSrc} alt={`${label} Logo`} />
+				<img class={size} src={logoSrc} alt={`${label} Logo`} loading="lazy" />
 			</div>
 		</div>
 	</div>