--- a/src/pages/_components/landing-page/Integrations.astro
+++ b/src/pages/_components/landing-page/Integrations.astro
@@ -73,13 +73,13 @@ import Vue from './integrations-tabs/Vue.astro';
 
 	<div class="my-4 sm:my-8 w-full border-t border-t-astro-gray-600"></div>
 
-	<ul class="mx-auto w-full max-w-screen-xl px-4 sm:px-8">
+	<div class="mx-auto w-full max-w-screen-xl px-4 sm:px-8">
 		<React />
 		<Vue />
 		<Preact />
 		<Svelte />
 		<Solid />
-	</ul>
+	</div>
 </section>
 
 <script>