--- a/src/pages/_components/landing-page/IslandGraphic.astro
+++ b/src/pages/_components/landing-page/IslandGraphic.astro
@@ -17,7 +17,7 @@ const { label, score, size = 'size-5', src, glow } = Astro.props;
 			style={{ width: `${score}%` }}
 			class="px-2 h-full flex items-center gap-2 bg-astro-gray-500 group-data-[glow]:bg-blue-green-gradient rounded-lg"
 		>
-			<img class={`${size}`} src={src} alt={`${label} logo`} />
+			<img class={`${size}`} src={src} alt={`${label} logo`} loading="lazy" />
 		</div>
 		<p class="sr-only">{label} Core Web Vitals Passing</p>
 		<p