--- a/src/pages/_components/landing-page/IslandGraphic.astro
+++ b/src/pages/_components/landing-page/IslandGraphic.astro
@@ -1,5 +1,15 @@
 ---
-const { label, score, size = 'size-5', src, glow } = Astro.props;
+import Icon from 'astro-icon';
+
+interface Props {
+	label: string;
+	score: string;
+	size?: `size-${number}` | `size-[${string}]`;
+	icon: string;
+	glow?: boolean;
+}
+
+const { label, score, size = 'size-5', icon, glow } = Astro.props;
 ---
 
 <div
@@ -17,7 +27,7 @@ const { label, score, size = 'size-5', src, glow } = Astro.props;
 			style={{ width: `${score}%` }}
 			class="px-2 h-full flex items-center gap-2 bg-astro-gray-500 group-data-[glow]:bg-blue-green-gradient rounded-lg"
 		>
-			<img class={`${size}`} src={src} alt={`${label} logo`} loading="lazy" />
+			<Icon name={icon} class={size} aria-label={`${label} logo`} />
 		</div>
 		<p class="sr-only">{label} Core Web Vitals Passing</p>
 		<p