--- a/src/pages/_components/landing-page/Islands.astro
+++ b/src/pages/_components/landing-page/Islands.astro
@@ -21,22 +21,11 @@ import IslandGraphic from './IslandGraphic.astro';
 		<p class="text-astro-gray-300">Real-World Core Web Vitals</p>
 
 		<div class="mt-8 mb-5">
-			<IslandGraphic
-				label="Astro"
-				score="60"
-				size="size-[22px]"
-				src="/assets/press/astro-icon-light.svg"
-				glow
-			/>
-			<IslandGraphic label="WordPress" score="38" src="/logos/wordpress.svg" />
-			<IslandGraphic
-				label="Gatsby"
-				score="38"
-				size="size-[19px]"
-				src="/assets/web-vitals/gatsby-logo.svg"
-			/>
-			<IslandGraphic label="Next.js" score="24" src="/assets/web-vitals/next-js-logo.svg" />
-			<IslandGraphic label="Nuxt" score="20" size="size-6" src="/assets/web-vitals/nuxt-logo.svg" />
+			<IslandGraphic label="Astro" score="60" size="size-[22px]" icon="logos/astro" glow />
+			<IslandGraphic label="WordPress" score="38" icon="logos/wordpress" />
+			<IslandGraphic label="Gatsby" score="38" size="size-[19px]" icon="logos/gatsby" />
+			<IslandGraphic label="Next.js" score="24" icon="logos/next-js" />
+			<IslandGraphic label="Nuxt" score="20" size="size-6" icon="logos/nuxt" />
 		</div>
 
 		<p class="text-astro-gray-300 text-balance">