--- a/src/pages/_components/landing-page/ecosystem-tabs/Blogs.astro
+++ b/src/pages/_components/landing-page/ecosystem-tabs/Blogs.astro
@@ -9,12 +9,12 @@ import openblog from '../../../_assets/themes/openblog.png';
 import stablo from '../../../_assets/themes/stablo.png';
 import ThemeContainer from './ThemeContainer.astro';
 
-const astroNanoTheme = await getImage({ src: astroNano, format: 'webp', width: 1280 });
-const brutalTheme = await getImage({ src: brutal, format: 'webp', width: 1280 });
-const danteTheme = await getImage({ src: dante, format: 'webp', width: 1280 });
-const microblogTheme = await getImage({ src: microblog, format: 'webp', width: 1280 });
-const openblogTheme = await getImage({ src: openblog, format: 'webp', width: 1280 });
-const stabloTheme = await getImage({ src: stablo, format: 'webp', width: 1280 });
+const astroNanoTheme = await getImage({ src: astroNano, format: 'webp', width: 800 });
+const brutalTheme = await getImage({ src: brutal, format: 'webp', width: 800 });
+const danteTheme = await getImage({ src: dante, format: 'webp', width: 800 });
+const microblogTheme = await getImage({ src: microblog, format: 'webp', width: 800 });
+const openblogTheme = await getImage({ src: openblog, format: 'webp', width: 800 });
+const stabloTheme = await getImage({ src: stablo, format: 'webp', width: 800 });
 
 const themes = [
 	{