--- a/src/pages/_components/landing-page/ecosystem-tabs/Docs.astro
+++ b/src/pages/_components/landing-page/ecosystem-tabs/Docs.astro
@@ -6,9 +6,9 @@ import ion from '../../../_assets/themes/ion.png';
 import starlight from '../../../_assets/themes/starlight.png';
 import ThemeContainer from './ThemeContainer.astro';
 
-const ionTheme = await getImage({ src: ion, format: 'webp', width: 1280 });
-const astroDSTheme = await getImage({ src: astroDS, format: 'webp', width: 1280 });
-const starlightTheme = await getImage({ src: starlight, format: 'webp', width: 1280 });
+const ionTheme = await getImage({ src: ion, format: 'webp', width: 800 });
+const astroDSTheme = await getImage({ src: astroDS, format: 'webp', width: 800 });
+const starlightTheme = await getImage({ src: starlight, format: 'webp', width: 800 });
 
 const themes = [
 	{