--- a/src/pages/_components/landing-page/ecosystem-tabs/Ecommerce.astro
+++ b/src/pages/_components/landing-page/ecosystem-tabs/Ecommerce.astro
@@ -6,9 +6,9 @@ import astroShopify from '../../../_assets/themes/astro-shopify.png';
 import quickStore from '../../../_assets/themes/quick-store.png';
 import ThemeContainer from './ThemeContainer.astro';
 
-const skincareTheme = await getImage({ src: SKNCRE, format: 'webp', width: 1280 });
-const quickStoreTheme = await getImage({ src: quickStore, format: 'webp', width: 1280 });
-const astroShopifyTheme = await getImage({ src: astroShopify, format: 'webp', width: 1280 });
+const skincareTheme = await getImage({ src: SKNCRE, format: 'webp', width: 800 });
+const quickStoreTheme = await getImage({ src: quickStore, format: 'webp', width: 800 });
+const astroShopifyTheme = await getImage({ src: astroShopify, format: 'webp', width: 800 });
 
 const themes = [
 	{