--- a/src/pages/_components/landing-page/ecosystem-tabs/Portfolios.astro
+++ b/src/pages/_components/landing-page/ecosystem-tabs/Portfolios.astro
@@ -10,15 +10,15 @@ import swissfolio from '../../../_assets/themes/swissfolio.png';
 import Link from '../../Link.astro';
 import ThemeContainer from './ThemeContainer.astro';
 
-const astrofyTheme = await getImage({ src: astrofy, format: 'webp', width: 1280 });
-const guridoTheme = await getImage({ src: gurido, format: 'webp', width: 1280 });
-const esqueleteCVTheme = await getImage({ src: esqueleteCV, format: 'webp', width: 1280 });
-const shareLinksTheme = await getImage({ src: shareLinks, format: 'webp', width: 1280 });
-const swissfolioTheme = await getImage({ src: swissfolio, format: 'webp', width: 1280 });
+const astrofyTheme = await getImage({ src: astrofy, format: 'webp', width: 800 });
+const guridoTheme = await getImage({ src: gurido, format: 'webp', width: 800 });
+const esqueleteCVTheme = await getImage({ src: esqueleteCV, format: 'webp', width: 800 });
+const shareLinksTheme = await getImage({ src: shareLinks, format: 'webp', width: 800 });
+const swissfolioTheme = await getImage({ src: swissfolio, format: 'webp', width: 800 });
 const moreThemesPlaceholder = await getImage({
 	src: morePortfolioThemes,
 	format: 'webp',
-	width: 1280,
+	width: 800,
 });
 
 const themes = [