--- a/src/pages/_components/landing-page/ecosystem-tabs/Trending.astro
+++ b/src/pages/_components/landing-page/ecosystem-tabs/Trending.astro
@@ -60,6 +60,8 @@ const themes = [
 	{themes.map((theme) => <ThemeContainer theme={theme} />)}
 	<div class="size-fit group relative">
 		<img
+			loading="lazy"
+			decoding="async"
 			class="size-full rounded-xl isolate select-none"
 			src={moreThemesPlaceholder.src}
 			alt="An image showing all types of content driven websites you can build with Astro"