--- a/src/pages/_components/landing-page/ecosystem-tabs/Trending.astro
+++ b/src/pages/_components/landing-page/ecosystem-tabs/Trending.astro
@@ -10,12 +10,12 @@ import tailcast from '../../../_assets/themes/tailcast.png';
 import Link from '../../Link.astro';
 import ThemeContainer from './ThemeContainer.astro';
 
-const astrolusTheme = await getImage({ src: astrolus, format: 'webp', width: 1280 });
-const tailcastTheme = await getImage({ src: tailcast, format: 'webp', width: 1280 });
-const astroShipTheme = await getImage({ src: astroShip, format: 'webp', width: 1280 });
-const starlightTheme = await getImage({ src: starlight, format: 'webp', width: 1280 });
-const primaPersonaTheme = await getImage({ src: primaPersona, format: 'webp', width: 1280 });
-const moreThemesPlaceholder = await getImage({ src: moreThemes, format: 'webp', width: 1280 });
+const astrolusTheme = await getImage({ src: astrolus, format: 'webp', width: 800 });
+const tailcastTheme = await getImage({ src: tailcast, format: 'webp', width: 800 });
+const astroShipTheme = await getImage({ src: astroShip, format: 'webp', width: 800 });
+const starlightTheme = await getImage({ src: starlight, format: 'webp', width: 800 });
+const primaPersonaTheme = await getImage({ src: primaPersona, format: 'webp', width: 800 });
+const moreThemesPlaceholder = await getImage({ src: moreThemes, format: 'webp', width: 800 });
 
 const themes = [
 	{