--- a/src/pages/_components/landing-page/integrations-tabs/InterfaceExample.astro
+++ b/src/pages/_components/landing-page/integrations-tabs/InterfaceExample.astro
@@ -14,6 +14,8 @@ const { brandColor, highContrast = false } = Astro.props;
 				class="w-full h-40 md:h-36 lg:h-40 xl:h-56 rounded-xl object-center object-contain"
 				src={astroCapSwag.src}
 				alt="Official Astro Cap Swag"
+				loading="lazy"
+				decoding="async"
 			/>
 			<div class="flex items-center justify-between text-sm xl:text-base">
 				<h2 class="text-astro-gray-300 font-medium">Snapback Cap</h2>