--- a/src/pages/db/_components/Intro.astro
+++ b/src/pages/db/_components/Intro.astro
@@ -11,7 +11,7 @@ import IntroPoint from './IntroPoint.astro';
 		<strong>Astro DB</strong> is a fully managed SQL database designed exclusively for Astro.
 	</h2>
 
-	<dl class="grid grid-cols-1 gap-8 sm:grid-cols-2 md:gap-12 lg:grid-cols-3">
+	<div class="grid grid-cols-1 gap-8 sm:grid-cols-2 md:gap-12 lg:grid-cols-3">
 		<IntroPoint bgGradient="bg-red-pink-gradient">
 			<Fragment slot="icon">
 				<div
@@ -53,7 +53,7 @@ import IntroPoint from './IntroPoint.astro';
 				Everything is setup for you, like a TypeScript ORM with full type-safety out of the box.
 			</Fragment>
 		</IntroPoint>
-	</dl>
+	</div>
 </section>
 
 <style>