--- a/src/pages/db/_components/IntroPoint.astro
+++ b/src/pages/db/_components/IntroPoint.astro
@@ -2,7 +2,7 @@
 const { bgGradient } = Astro.props;
 ---
 
-<div>
+<article>
 	<div class={`h-fit w-fit rounded-full p-0.5 ${bgGradient}`}>
 		<div class="rounded-full bg-black">
 			<slot name="icon" />
@@ -12,4 +12,4 @@ const { bgGradient } = Astro.props;
 	<p class="text-balance body mt-1 text-astro-gray-200">
 		<slot name="description" />
 	</p>
-</div>
+</article>