--- a/src/pages/db/_components/Pill.astro
+++ b/src/pages/db/_components/Pill.astro
@@ -8,7 +8,7 @@ type Props = {
 const { title, subtitle, class: className } = Astro.props;
 ---
 
-<aside
+<div
 	class:list={[
 		'relative flex gap-2 rounded-full text-xs transition-colors duration-150 md:text-sm',
 		className,
@@ -18,10 +18,10 @@ const { title, subtitle, class: className } = Astro.props;
 		<span class="pill relative m-px rounded-full px-3 py-1">{title}</span>
 		{subtitle && <span>{subtitle}</span>}
 	</span>
-</aside>
+</div>
 
 <style>
-	aside::before {
+	div::before {
 		@apply absolute inset-0 border border-white/40;
 		content: '';
 		border-radius: inherit;