--- a/src/pages/db/_components/Subnav.astro
+++ b/src/pages/db/_components/Subnav.astro
@@ -5,7 +5,7 @@ import Badge from './Badge.astro';
 ---
 
 <nav
-	aria-label="Secondary"
+	aria-label="DB Sections"
 	class:list={[
 		'container fixed left-0 right-0 top-0 z-20 flex hidden h-header items-center bg-black shadow-lg sm:flex',
 		Astro.props.class,