… public/assets/web-vitals/gatsby-logo.svg
… public/assets/web-vitals/next-js-logo.svg
… public/assets/web-vitals/nuxt-logo.svg
… public/assets/web-vitals/wordpress-logo.svg
… src/content/pages/privacy.md
… src/content/pages/terms.md
… src/icons/logos/astro.svg
… src/icons/logos/gatsby.svg
… src/icons/logos/nuxt.svg
… src/icons/logos/wordpress.svg
… src/pages/db/_components/UseCaseTab.astro
… src/pages/db/_components/UseCaseTabs.astro
… src/pages/db/_components/features/ORM.astro
… src/pages/db/_components/features/SQLCommands.astro
… src/pages/press/_components/PressAssetCard.astro
… src/pages/themes/_components/ThemeCard.astro
… src/pages/wallpapers/_components/WallpaperAssetCard.astro