This is a bit of an omnibus PR tackling many small paper cuts around accessibility and performance on astro.build. That makes it a little tricky to review, but here are the changes broken down by where they were made. For the most part nothing has changed visually, or if it has, in very tiny ways I would expect no-one to notice — I’ve marked these with 👁️.

#### Miscellaneous

- Adds missing accessible labels to UI and images in the gallery components for individual theme pages, e.g. [/themes/details/portfolio/](https://deploy-preview-1228--astro-www-2.netlify.app/themes/details/portfolio/)
- Fixes heading levels in some pages that were jumping `h1` => `h3`, e.g. [/privacy/](https://deploy-preview-1228--astro-www-2.netlify.app/privacy/)
- Fixes some card grid usage where the structure was `ul > article`. These are now `ul > li > article`, e.g. [/press/](https://deploy-preview-1228--astro-www-2.netlify.app/press/#assets)
- Uses trailing slashes on some internal theme and showcase links to match canonical URLs (showcase pages were redirecting, so this change improves performance; theme pages were not so this improves SEO)

#### `/db/` landing page

- Fixes [use-case tabs on `/db/`](https://deploy-preview-1228--astro-www-2.netlify.app/db/#use-cases) to address incorrect `aria-*` attributes (also reduces JS quite a bit because the custom element was managing a bunch of unnecessary attribute state).
- Fixes text contrast in [`/db/` features section](https://deploy-preview-1228--astro-www-2.netlify.app/db/#features). (👁️ This is a visual change that does make the illustrations impacted slightly less subtle, but text contrast requirements still need to be met.)
- Fixes accessible labelling of secondary navbar in `/db/`.
- Fixes miscellaneous misused elements and attributes in `/db/` (e.g. `<dl>` used without `<dt>` and `<dd>`)

#### Homepage

- Uses `div > section[role="tabpanel"]` for the tabbed interfaces on the landing page instead of `ul > li[role="tabpanel"]`. The `tabpanel` role removes the list-item semantics from an `li` making it invalid to use inside a `ul`.
- Reduces the amount of blur used on background image in the hero section of the landing page. Safari in particular was really struggling to render the 64px blur over such a large area. Reducing the blur amount has minimal visual impact while fixing performance.
- Sets `loading="lazy"` on images below the fold on the landing page to improve page load performance.
- Uses an `<img>` instead of a background image for the background of the homepage hero element to prioritise its loading and improve perceived page performance.
	- 👁️ I also opted to change how the background works across viewports. Currently it is anchored to the left meaning when the viewport gets narrower only the black-to-blue portion of the image is used, which felt like a mistake. It now holds its position centrally so that the mobile view is a more evenly distributed texture similar to the desktop view.
- Reduces size of overly large images on homepage to match rendered size better.
- Uses inline SVGs for logos in islands performance chart instead of `<img>` elements.
- Fixes content shift from missing `width` and `height` attributes in “ecosystem” themes tabs on homepage.

##### Homepage headline numbers

| | Before | After | Change |
|---|---|---|---|
| Initial page weight (uncompressed)<sup>*</sup> | 1,902 kB | 617 kB | -68% |
| Initial page weight (compressed)<sup>*</sup> | 1,561 kB | 275 kB | -82% |
| LCP (first visit, mobile, 4G) | 2.4s | 1.9s | -20% |
| LCP (repeat visit, mobile, 4G) | 2.4s | 1.6s | -33% |
| LCP (first visit, desktop, cable) | 0.9s | 0.8s | -11% |
| LCP (repeat visit, desktop, cable) | 1s | 0.6s | -40% |
| Visually complete (desktop, cable)<sup>†</sup> | 2.8s | 0.9s | -68% |

<sup>* This is massively reduced thanks to lazy loading assets not needed for the initial view</sup>
<sup>† The visually complete metric matches when the hero background currently loads. While on a reasonably fast desktop connection LCP is not improved so much (still depends on font loading), the background image now loads way faster.</sup>

Web vitals numbers are from WebPageTest.com. Full runs:
- Before: [Mobile, 4G](https://www.webpagetest.org/result/240823_AiDcJH_5A1/); [Desktop, cable](https://www.webpagetest.org/result/240823_BiDcAC_69Z/)
- After: [Mobile, 4G](https://www.webpagetest.org/result/240823_BiDc4Z_5BT/); [Desktop, cable](https://www.webpagetest.org/result/240823_AiDcRE_6A8/)
- Comparison: [Mobile, 4G](https://www.webpagetest.org/video/compare.php?tests=240823_AiDcJH_5A1,240823_BiDc4Z_5BT); [Desktop, cable](https://www.webpagetest.org/video/compare.php?tests=240823_AiDcRE_6A8,240823_BiDcAC_69Z)

Here are side-by-side loading comparison for the before/after runs:

https://github.com/user-attachments/assets/e2b5c28c-87c5-40c7-a6c3-428fb4181cf5

https://github.com/user-attachments/assets/e52c4863-cbb7-46f6-9552-ee0320bf3b92




## Browser Test Checklist

<!--
Test on as many of these browsers as you can, ideally 3+ of them.

Tests for all browsers are **strongly recommended** if this PR includes:

- Large gradients or the usage of `mask-image`, which can cause perf issues on Firefox Android (https://github.com/withastro/astro.build/pull/780)
- Font changes, which can behave strangely on Safari (https://github.com/withastro/astro.build/pull/1028)
- Adding SVGs, which can behave strangely on Safari (https://github.com/withastro/astro.build/pull/769)
-->

I have tested this PR on at least three of the following browsers:

- [x] Chrome / Chromium
- [x] Firefox
- [x] Android Firefox (the /db/ page's performance is still diabolically awful here but that's not new to this PR)
- [x] Safari
- [ ] iOS Safari